% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_int.R
\name{lorenz_int}
\alias{lorenz_int}
\title{Computes income inequality statistics derived with Lorenz interpolation.}
\usage{
lorenz_int(freqs, bounds, G, stat = "gini", slope_parm = 0.9)
}
\arguments{
\item{freqs}{A vector of counts in income brackets.}

\item{bounds}{A vector of income bracket boundaries.}

\item{G}{Grand mean of income distribution.}

\item{stat}{(optional) Return income statistic instead of sample incomes.
#' @param slope_parm (default = .9) Slope parameter that influences the shape of the function fitted to the Lorenz curve.}
}
\value{
Income inequality statistics derived with Lorenz interpolation.
}
\description{
Computes income inequality statistics derived with Lorenz interpolation.
}
\examples{
ex_freqs <- c(45, 31, 33, 27, 43, 40, 51, 50, 63, 97, 121, 132, 64, 54, 32, 12)
ex_bounds <- c(0, 10000, 15000, 20000, 25000, 30000, 35000, 40000, 45000, 50000, 60000, 75000,
100000, 125000, 150000, 200000)
ex_mean <- 66500
lorenz_int(ex_freqs, ex_bounds, ex_mean)
}
