% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcib.R
\name{mcib}
\alias{mcib}
\title{Takes number of draws, freqs, bounds, and aggregate for region and returns sample of incomes.}
\usage{
mcib(freqs, bounds, mean)
}
\arguments{
\item{freqs}{A vector of counts in income brackets.}

\item{bounds}{A vector of income bracket boundaries.}

\item{mean}{Grand mean of income distribution.}
}
\value{
Income inequality statistics derived with mean-constrained integration over brackets.
}
\description{
Takes number of draws, freqs, bounds, and aggregate for region and returns sample of incomes.
}
\examples{
bracket_frequencies <- c(45, 31, 33, 27, 43, 40, 51, 50, 63, 97, 121, 132, 64, 54, 32, 12)
bracket_boundaries <- c(0, 10000, 15000, 20000, 25000, 30000, 35000, 40000, 45000, 50000,
 60000, 75000, 100000, 125000, 150000, 200000)
mean <- 66500
mcib(bracket_frequencies, bracket_boundaries, mean)
}
